Added level two tilemap and imported Material Editor
There's a z-order issue when using two tilemaps as You can only sort the player within one. The most correct approach would be to make the top half and bottom half of the character into separate sprites. This may be too much work or it may actually make it easier as we can divide the walk cycle and the attack animations